feat: create ClientProviders component for centralized theming and context providers
